+++ bindings/angular2/examples/select.spec.js
@@ -8,4 +8,11 @@ describe('select.html', () => {
   it('should have ons-select elements', () => {
     expect($('ons-select').isPresent()).toBeTruthy();
   });
+
+  describe('reactive form', () => {
+    it('should change form value when the selection is changed', () => {
+      $('#animal-3').click();
+      expect($('#selected-animal').getText()).toBe('Elephant');
+    });
+  });
 });