+++ core/src/elements/base/base-input.js
@@ -77,6 +77,7 @@ export default class BaseInputElement extends BaseElement {
     this.appendChild(util.createFragment(this._template));
 
     this._setInputId();
+
     this._updateBoundAttributes();
 
     ModifierUtil.initModifier(this, this._scheme);